requirements.txt
google
python-dotenv
discord.py
beautifulsoup4==4.8.1
celery==4.4.0
decorator==4.4.1
Django==3.0
django-allauth
django-bootstrap4==1.1.1
django-jsonify==0.3.0
importlib-metadata==1.3.0
oauthlib==3.1.0
psycopg2-binary
python3-openid==3.1.0
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
wrapt==1.11.2
